Analysis of the program
The program boots up very nicely and puts the users right into the app ready to go. The parts that have been finished look good and user friendly. The unit tests are very well done but I saw some edge cases unaccounted for.

Was the program available in UC Github on time?

Yes, I was given access to the repo immediately.

Is the program documented/commented well enough for you to understand?

The program is well organized and documented very well. It was very easy for me to understand what it was doing.

Does the program compile?

Yes, the app launched with no fatal errors.

Rationale behind your changes.

While going through the app I noticed that the unit cases do not check for the creation of bad ListItems. I added the tests to do so and then added the verification in the ListItems class to throw errors when a bad entry is given. I also removed a few spots where old code was commented out. It is best to use version control to keep unused or bad code out of the master as apposed to comments.

Three Technical Concepts I learned

I learned more about recylerviews and how to use them. They get tricky when adding and removing things and they handle it well.
I learned about adding items to firebase which is helpful as my group is currently having issues with ours.
I learned more about adding LiveData to a feed which is helpful as that is next on my group's TODO list.