Past List Functionality
Added some increased functionality to the past list tab. Now it can view the list names on the spinner, but the names themselves do not yet pop up - likely an issue with translating the strings across a database and multiple classes. The empty values, however, stay up-to-date with livedata, and increase or decrease depending on the number of lists.